Shuffled meerjarenplan aanpassing
